# api-docs.json
{"openapi":"3.1.0","info":{"title":"Employee API","description":"API documentation for managing employees","version":"1.0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"paths":{"/api/v1/employee/{id}":{"get":{"tags":["employee-controller"],"operationId":"getStudentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDto"}}}}}},"put":{"tags":["employee-controller"],"operationId":"updateEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeDto"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDto"}}}}}},"delete":{"tags":["employee-controller"],"operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"200":{"description":"OK"}}},"patch":{"tags":["employee-controller"],"operationId":"updateUsingPatch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDto"}}}}}}},"/api/v1/employee":{"get":{"tags":["employee-controller"],"operationId":"getAllStudent","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeDto"}}}}}}},"post":{"tags":["employee-controller"],"operationId":"AddEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeDto"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDto"}}}}}}},"/api/v1/employee/{id}/photo":{"post":{"tags":["employee-controller"],"operationId":"uploadEmployeePhoto","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/v1/employee/photo/{filename}":{"get":{"tags":["employee-controller"],"operationId":"getPhoto","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}}},"components":{"schemas":{"EmployeeDto":{"type":"object","properties":{"employee_id":{"type":"string","format":"uuid"},"name":{"type":"string","maxLength":100,"minLength":2},"email":{"type":"string","format":"email","maxLength":50,"minLength":7},"phone_number":{"type":"string","maxLength":10,"minLength":10},"department":{"type":"string","maxLength":50,"minLength":2},"role":{"type":"string","maxLength":50,"minLength":2},"city":{"type":"string","maxLength":50,"minLength":2},"address":{"type":"string"},"photo":{"type":"string"}},"required":["city","department","email","name","phone_number","role"]}}}}